+++ src/UI/TransactionInfo/components/transactionInfoFields/TransactionInfoGasPrice/TransactionInfoGasPrice.tsx
@@ -3,15 +3,19 @@ import React from 'react';
 import { N_A } from 'constants/index';
 import { FormatAmount } from 'UI/FormatAmount/FormatAmount';
 
-import { WithTransactionType } from '../../../../../UI/types';
+import {
+  WithTransactionType,
+  WithClassnameType
+} from '../../../../../UI/types';
 import { DetailItem } from '../../DetailItem';
 
 import styles from './styles.scss';
 
 export const TransactionInfoGasUsed = ({
+  className,
   transaction
-}: WithTransactionType) => (
-  <DetailItem title='Gas Price'>
+}: WithTransactionType & WithClassnameType) => (
+  <DetailItem className={className} title='Gas Price'>
     {transaction.gasPrice != null ? (
       <FormatAmount
         value={transaction.gasPrice.toString()}