+++ src/UI/TransactionInfo/components/transactionInfoFields/TransactionInfoGasUsed/TransactionInfoGasUsed.tsx
@@ -2,15 +2,19 @@ import React from 'react';
 
 import { N_A } from 'constants/index';
 
-import { WithTransactionType } from '../../../../../UI/types';
+import {
+  WithTransactionType,
+  WithClassnameType
+} from '../../../../../UI/types';
 import { DetailItem } from '../../DetailItem';
 
 import styles from './styles.scss';
 
 export const TransactionInfoGasPrice = ({
+  className,
   transaction
-}: WithTransactionType) => (
-  <DetailItem title='Gas Used'>
+}: WithTransactionType & WithClassnameType) => (
+  <DetailItem className={className} title='Gas Used'>
     {transaction.gasUsed != null ? (
       transaction.gasUsed.toLocaleString('en')
     ) : (