+++ src/UI/TransactionInfo/components/transactionInfoFields/TransactionInfoHash/TransactionInfoHash.tsx
@@ -1,14 +1,19 @@
 import React from 'react';
-
 import { CopyButton } from 'UI/CopyButton';
 
-import { WithTransactionType } from '../../../../../UI/types';
+import {
+  WithClassnameType,
+  WithTransactionType
+} from '../../../../../UI/types';
 import { DetailItem } from '../../DetailItem';
 
 import styles from './styles.scss';
 
-export const TransactionInfoHash = ({ transaction }: WithTransactionType) => (
-  <DetailItem title='Hash'>
+export const TransactionInfoHash = ({
+  className,
+  transaction
+}: WithTransactionType & WithClassnameType) => (
+  <DetailItem className={className} title='Hash'>
     <div className={styles.hash}>
       {transaction.txHash}