+++ src/UI/TransactionsTable/components/TransactionDirectionBadge.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import classNames from 'classnames';
 
+import globalStyles from 'assets/sass/main.scss';
 import { WithTransactionType } from '../../../UI/types';
 
 import styles from './transactionsTable.styles.scss';
-import globalStyles from 'assets/sass/main.scss';
 
 export const TransactionDirectionBadge = ({
   transaction